Fix post edit navigation by restructuring routes
